test: update COM tests
