+++ src/Odotocodot.OneNote.Linq.Tests/Odotocodot.OneNote.Linq.Tests.csproj
@@ -14,6 +14,7 @@
 	  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
 	</PackageReference>
 	<PackageReference Include="NUnit3TestAdapter" Version="6.0.0" />
+	<PackageReference Include="Shouldly" Version="4.3.0" />
   </ItemGroup>
 
   <ItemGroup>
